jshint + refactor + modal focus FIX + delete disabled FIX

+++ rest-api/src/main/webapp/app/schedule/courses/courses.js
@@ -41,8 +41,10 @@ define([
         ];
         $scope.hours = [];
         $scope.minutes = [];
-        for (var i = 7; i < 24; i++) {$scope.hours.push("" + i)}
-        for (var i = 0; i < 60; i+=5) {
+        for (var i = 7; i < 24; i++) {
+            $scope.hours.push("" + i);
+        }
+        for (i = 0; i < 60; i += 5) {
             var j = i;
             if (j < 10) {
                 j = "0" + j;
@@ -55,7 +57,7 @@ define([
             column: 'startTime',
             descending: false
         };
-        $scope.toggleSort = function(column) {
+        $scope.toggleSort = function (column) {
             var sort = $scope.sort;
             if (sort.column == column) {
                 sort.descending = !sort.descending;
@@ -70,7 +72,7 @@ define([
                 return sort.descending ? "fa fa-caret-down" : "fa fa-caret-up";
             }
             return "";
-        }
+        };
 
         // =======================================
         $scope.add = function () {
@@ -89,7 +91,7 @@ define([
             });
         };
         $scope.update = function (course) {
-            delete course['edit'];
+            delete course.edit;
             courseHttpClient.update({ sid: course.sid }, course).$promise.then(
                 function () {
                     notificationService.success("Pomyślnie zapisano");
@@ -99,7 +101,7 @@ define([
                         }
                     );
                 });
-        }
+        };
 
         $scope.publish = function (course) {
             courseHttpClient.publish({ sid: course.sid }).$promise.then(
@@ -111,32 +113,34 @@ define([
                         });
                     notificationService.success("Pomyślnie zapisano");
                 });
-        }
+        };
 
         $scope.setDay = function (day) {
-            $scope.selected = null;
+            $scope.selected = undefined;
             $scope.day = day;
-        }
+        };
 
         $scope.delete = function (course) {
-            var modalInstance = $modal.open({
-                templateUrl: 'app/core/modal/deleteConfirm.html',
-                controller: "deleteConfirmDialogController"
-            });
+            if (course !== undefined) {
+                var modalInstance = $modal.open({
+                    templateUrl: 'app/core/modal/deleteConfirm.html',
+                    controller: "deleteConfirmDialogController"
+                });
 
-            modalInstance.result.then(function () {
-                $scope.selected = null;
-                courseHttpClient.delete({sid: course.sid}).$promise.then(
-                    function () {
-                        notificationService.success("Pomyślnie usunięto");
-                        courseHttpClient.findAll().$promise.then(
-                            function (result) {
-                                $scope.classes = result;
-                            })
-                    }
-                )
-            });
-        }
+                modalInstance.result.then(function () {
+                    $scope.selected = undefined;
+                    courseHttpClient.delete({sid: course.sid}).$promise.then(
+                        function () {
+                            notificationService.success("Pomyślnie usunięto");
+                            courseHttpClient.findAll().$promise.then(
+                                function (result) {
+                                    $scope.classes = result;
+                                });
+                        }
+                    );
+                });
+            }
+        };
 
         $scope.deactivate = function (course) {
             courseHttpClient.deactivate({ sid: course.sid }).$promise.then(
@@ -148,11 +152,11 @@ define([
                         });
                     notificationService.success("Pomyślnie zapisano");
                 });
-        }
+        };
 
         $scope.select = function (course) {
-            if ($scope.selected != null && $scope.selected.sid == course.sid) {
-                $scope.selected = null;
+            if ($scope.selected !== undefined && $scope.selected.sid == course.sid) {
+                $scope.selected = undefined;
                 $scope.lessons = [];
                 return;
             }
@@ -166,10 +170,10 @@ define([
                     saving: false,
                     hover: false,
                     oldVal: {}
-                }
+                };
             }
 
-        }
+        };
         $scope.edit = function (object, property) {
             object[property].edit = true;
             object[property].oldVal = object[property].value;
@@ -178,22 +182,20 @@ define([
                 object[property].value = {
                     hours: obj[0],
                     minutes: obj[1]
-                }
+                };
             }
-
-
-        }
+        };
         $scope.cancel = function (object, property) {
             object[property].value = object[property].oldVal;
             object[property].edit = false;
             object[property].hover = false;
-        }
+        };
         $scope.hover = function (object, property) {
             object[property].hover = true;
-        }
+        };
         $scope.leave = function (object, property) {
             object[property].hover = false;
-        }
+        };
         $scope.save = function (course, property) {
             course[property].saving = true;
             if (property == STYLE_KEY) {
@@ -224,7 +226,7 @@ define([
             }
             else {
                 var obj = _.findWhere($scope.classes, {sid: course.sid});
-                if (obj != null) {
+                if (obj !== undefined) {
                     var newVal = course[property].value;
                     if (property == START_TIME_KEY || property == END_TIME_KEY) {
                         obj[property] = newVal.hours + ":" + newVal.minutes;
@@ -246,6 +248,6 @@ define([
                         });
                 }
             }
-        }
+        };
     }]);
 });
