+++ rest-api/src/main/webapp/app/tickets/typeGroups/ticketTypeGroups.js
@@ -22,7 +22,7 @@ define([
             column: 'name',
             descending: false
         };
-        $scope.toggleSort = function(column) {
+        $scope.toggleSort = function (column) {
             var sort = $scope.sort;
             if (sort.column == column) {
                 sort.descending = !sort.descending;
@@ -37,7 +37,7 @@ define([
                 return sort.descending ? "fa fa-caret-down" : "fa fa-caret-up";
             }
             return "";
-        }
+        };
 
         // =======================================
         $scope.add = function () {
@@ -53,11 +53,11 @@ define([
                         $scope.groups = result;
                     });
             });
-        }
+        };
         // =======================================
         $scope.select = function (group) {
-            if ($scope.selected != null && $scope.selected.sid == group.sid) {
-                $scope.selected = null;
+            if ($scope.selected !== undefined && $scope.selected.sid == group.sid) {
+                $scope.selected = undefined;
                 return;
             }
             $scope.selected = angular.copy(group);
@@ -69,30 +69,30 @@ define([
                     saving: false,
                     hover: false,
                     oldVal: {}
-                }
+                };
             }
-        }
+        };
         $scope.edit = function (object, property) {
             object[property].edit = true;
             object[property].oldVal = object[property].value;
-        }
+        };
         $scope.cancel = function (object, property) {
             object[property].value = object[property].oldVal;
             object[property].edit = false;
             object[property].hover = false;
-        }
+        };
         $scope.hover = function (object, property) {
             object[property].hover = true;
-        }
+        };
         $scope.leave = function (object, property) {
             object[property].hover = false;
-        }
+        };
         $scope.save = function (object, property) {
             object[property].saving = true;
             if (property == NAME_KEY) {
                 var obj = _.findWhere($scope.groups, {sid: object.sid});
                 obj[NAME_KEY] = object[NAME_KEY].value;
-                if (obj != null) {
+                if (obj !== undefined) {
                     ticketTypeGroupHttpClient.update({ sid: object.sid }, obj).$promise.then(
                         function () {
                             object[property].edit = false;
@@ -107,26 +107,28 @@ define([
                         });
                 }
             }
-        }
+        };
         // =======================================
         $scope.delete = function (group) {
-            var modalInstance = $modal.open({
-                templateUrl: 'app/core/modal/deleteConfirm.html',
-                controller: "deleteConfirmDialogController"
-            });
+            if (group !== undefined) {
+                var modalInstance = $modal.open({
+                    templateUrl: 'app/core/modal/deleteConfirm.html',
+                    controller: "deleteConfirmDialogController"
+                });
 
-            modalInstance.result.then(function () {
-                $scope.selected = null;
-                ticketTypeGroupHttpClient.delete({sid: group.sid}).$promise.then(
-                    function () {
-                        notificationService.success("Pomyślnie usunięto");
-                        ticketTypeGroupHttpClient.findAll().$promise.then(
-                            function (result) {
-                                $scope.groups = result;
-                            })
-                    }
-                )
-            });
-        }
+                modalInstance.result.then(function () {
+                    $scope.selected = undefined;
+                    ticketTypeGroupHttpClient.delete({sid: group.sid}).$promise.then(
+                        function () {
+                            notificationService.success("Pomyślnie usunięto");
+                            ticketTypeGroupHttpClient.findAll().$promise.then(
+                                function (result) {
+                                    $scope.groups = result;
+                                });
+                        }
+                    );
+                });
+            }
+        };
     }]);
 });
